+++ graph/api/api_test.go
@@ -22,6 +22,7 @@ import (
 	"k8s.io/client-go/tools/clientcmd/api"
 
 	"github.com/kiali/kiali/business"
+	"github.com/kiali/kiali/business/authentication"
 	"github.com/kiali/kiali/config"
 	"github.com/kiali/kiali/graph"
 	"github.com/kiali/kiali/kubernetes/kubetest"
@@ -1077,7 +1078,7 @@ func TestAppGraph(t *testing.T) {
 	mr := mux.NewRouter()
 	mr.HandleFunc("/api/namespaces/graph", http.HandlerFunc(
 		func(w http.ResponseWriter, r *http.Request) {
-			context := config.SetAuthInfoContext(r.Context(), &api.AuthInfo{Token: "test"})
+			context := authentication.SetAuthInfoContext(r.Context(), &api.AuthInfo{Token: "test"})
 			code, config := fut(context, nil, client, graph.NewOptions(r.WithContext(context)))
 			respond(w, code, config)
 		}))
@@ -1116,7 +1117,7 @@ func TestVersionedAppGraph(t *testing.T) {
 	mr := mux.NewRouter()
 	mr.HandleFunc("/api/namespaces/graph", http.HandlerFunc(
 		func(w http.ResponseWriter, r *http.Request) {
-			context := config.SetAuthInfoContext(r.Context(), &api.AuthInfo{Token: "test"})
+			context := authentication.SetAuthInfoContext(r.Context(), &api.AuthInfo{Token: "test"})
 			code, config := fut(context, nil, client, graph.NewOptions(r.WithContext(context)))
 			respond(w, code, config)
 		}))
@@ -1155,7 +1156,7 @@ func TestServiceGraph(t *testing.T) {
 	mr := mux.NewRouter()
 	mr.HandleFunc("/api/namespaces/graph", http.HandlerFunc(
 		func(w http.ResponseWriter, r *http.Request) {
-			context := config.SetAuthInfoContext(r.Context(), &api.AuthInfo{Token: "test"})
+			context := authentication.SetAuthInfoContext(r.Context(), &api.AuthInfo{Token: "test"})
 			code, config := fut(context, nil, client, graph.NewOptions(r.WithContext(context)))
 			respond(w, code, config)
 		}))
@@ -1194,7 +1195,7 @@ func TestWorkloadGraph(t *testing.T) {
 	mr := mux.NewRouter()
 	mr.HandleFunc("/api/namespaces/graph", http.HandlerFunc(
 		func(w http.ResponseWriter, r *http.Request) {
-			context := config.SetAuthInfoContext(r.Context(), &api.AuthInfo{Token: "test"})
+			context := authentication.SetAuthInfoContext(r.Context(), &api.AuthInfo{Token: "test"})
 			code, config := fut(context, nil, client, graph.NewOptions(r.WithContext(context)))
 			respond(w, code, config)
 		}))
@@ -1233,7 +1234,7 @@ func TestRatesGraphSent(t *testing.T) {
 	mr := mux.NewRouter()
 	mr.HandleFunc("/api/namespaces/graph", http.HandlerFunc(
 		func(w http.ResponseWriter, r *http.Request) {
-			context := config.SetAuthInfoContext(r.Context(), &api.AuthInfo{Token: "test"})
+			context := authentication.SetAuthInfoContext(r.Context(), &api.AuthInfo{Token: "test"})
 			code, config := fut(context, nil, client, graph.NewOptions(r.WithContext(context)))
 			respond(w, code, config)
 		}))
@@ -1272,7 +1273,7 @@ func TestRatesGraphReceived(t *testing.T) {
 	mr := mux.NewRouter()
 	mr.HandleFunc("/api/namespaces/graph", http.HandlerFunc(
 		func(w http.ResponseWriter, r *http.Request) {
-			context := config.SetAuthInfoContext(r.Context(), &api.AuthInfo{Token: "test"})
+			context := authentication.SetAuthInfoContext(r.Context(), &api.AuthInfo{Token: "test"})
 			code, config := fut(context, nil, client, graph.NewOptions(r.WithContext(context)))
 			respond(w, code, config)
 		}))
@@ -1311,7 +1312,7 @@ func TestRatesGraphTotal(t *testing.T) {
 	mr := mux.NewRouter()
 	mr.HandleFunc("/api/namespaces/graph", http.HandlerFunc(
 		func(w http.ResponseWriter, r *http.Request) {
-			context := config.SetAuthInfoContext(r.Context(), &api.AuthInfo{Token: "test"})
+			context := authentication.SetAuthInfoContext(r.Context(), &api.AuthInfo{Token: "test"})
 			code, config := fut(context, nil, client, graph.NewOptions(r.WithContext(context)))
 			respond(w, code, config)
 		}))
@@ -1350,7 +1351,7 @@ func TestRatesGraphNone(t *testing.T) {
 	mr := mux.NewRouter()
 	mr.HandleFunc("/api/namespaces/graph", http.HandlerFunc(
 		func(w http.ResponseWriter, r *http.Request) {
-			context := config.SetAuthInfoContext(r.Context(), &api.AuthInfo{Token: "test"})
+			context := authentication.SetAuthInfoContext(r.Context(), &api.AuthInfo{Token: "test"})
 			code, config := fut(context, nil, client, graph.NewOptions(r.WithContext(context)))
 			respond(w, code, config)
 		}))
@@ -1626,7 +1627,7 @@ func TestWorkloadNodeGraph(t *testing.T) {
 	mr := mux.NewRouter()
 	mr.HandleFunc("/api/namespaces/{namespace}/workloads/{workload}/graph", http.HandlerFunc(
 		func(w http.ResponseWriter, r *http.Request) {
-			context := config.SetAuthInfoContext(r.Context(), &api.AuthInfo{Token: "test"})
+			context := authentication.SetAuthInfoContext(r.Context(), &api.AuthInfo{Token: "test"})
 			code, config := fut(context, nil, client, graph.NewOptions(r.WithContext(context)))
 			respond(w, code, config)
 		}))
@@ -1902,7 +1903,7 @@ func TestAppNodeGraph(t *testing.T) {
 	mr := mux.NewRouter()
 	mr.HandleFunc("/api/namespaces/{namespace}/applications/{app}/graph", http.HandlerFunc(
 		func(w http.ResponseWriter, r *http.Request) {
-			context := config.SetAuthInfoContext(r.Context(), &api.AuthInfo{Token: "test"})
+			context := authentication.SetAuthInfoContext(r.Context(), &api.AuthInfo{Token: "test"})
 			code, config := fut(context, nil, client, graph.NewOptions(r.WithContext(context)))
 			respond(w, code, config)
 		}))
@@ -2178,7 +2179,7 @@ func TestVersionedAppNodeGraph(t *testing.T) {
 	mr := mux.NewRouter()
 	mr.HandleFunc("/api/namespaces/{namespace}/applications/{app}/versions/{version}/graph", http.HandlerFunc(
 		func(w http.ResponseWriter, r *http.Request) {
-			context := config.SetAuthInfoContext(r.Context(), &api.AuthInfo{Token: "test"})
+			context := authentication.SetAuthInfoContext(r.Context(), &api.AuthInfo{Token: "test"})
 			code, config := fut(context, nil, client, graph.NewOptions(r.WithContext(context)))
 			respond(w, code, config)
 		}))
@@ -2264,7 +2265,7 @@ func TestServiceNodeGraph(t *testing.T) {
 	mr := mux.NewRouter()
 	mr.HandleFunc("/api/namespaces/{namespace}/services/{service}/graph", http.HandlerFunc(
 		func(w http.ResponseWriter, r *http.Request) {
-			context := config.SetAuthInfoContext(r.Context(), &api.AuthInfo{Token: "test"})
+			context := authentication.SetAuthInfoContext(r.Context(), &api.AuthInfo{Token: "test"})
 			code, config := fut(context, nil, client, graph.NewOptions(r.WithContext(context)))
 			respond(w, code, config)
 		}))
@@ -2631,7 +2632,7 @@ func TestRatesNodeGraphTotal(t *testing.T) {
 	mr := mux.NewRouter()
 	mr.HandleFunc("/api/namespaces/{namespace}/workloads/{workload}/graph", http.HandlerFunc(
 		func(w http.ResponseWriter, r *http.Request) {
-			context := config.SetAuthInfoContext(r.Context(), &api.AuthInfo{Token: "test"})
+			context := authentication.SetAuthInfoContext(r.Context(), &api.AuthInfo{Token: "test"})
 			code, config := fut(context, nil, client, graph.NewOptions(r.WithContext(context)))
 			respond(w, code, config)
 		}))
@@ -3134,7 +3135,7 @@ func TestComplexGraph(t *testing.T) {
 	mr := mux.NewRouter()
 	mr.HandleFunc("/api/namespaces/graph", http.HandlerFunc(
 		func(w http.ResponseWriter, r *http.Request) {
-			context := config.SetAuthInfoContext(r.Context(), &api.AuthInfo{Token: "test"})
+			context := authentication.SetAuthInfoContext(r.Context(), &api.AuthInfo{Token: "test"})
 			code, config := fut(context, nil, client, graph.NewOptions(r.WithContext(context)))
 			respond(w, code, config)
 		}))
@@ -3421,7 +3422,7 @@ func TestMultiClusterSourceGraph(t *testing.T) {
 	mr := mux.NewRouter()
 	mr.HandleFunc("/api/namespaces/graph", http.HandlerFunc(
 		func(w http.ResponseWriter, r *http.Request) {
-			context := config.SetAuthInfoContext(r.Context(), &api.AuthInfo{Token: "test"})
+			context := authentication.SetAuthInfoContext(r.Context(), &api.AuthInfo{Token: "test"})
 			code, config := fut(context, nil, client, graph.NewOptions(r.WithContext(context)))
 			respond(w, code, config)
 		}))
