+++ handlers/authentication.go
@@ -74,7 +74,7 @@ func (aHandler AuthenticationHandler) Handle(next http.Handler) http.Handler {
 				http.Error(w, http.StatusText(http.StatusBadRequest), http.StatusBadRequest)
 				log.Errorf("No authInfo: %v", http.StatusBadRequest)
 			}
-			ctx := config.SetAuthInfoContext(r.Context(), authInfo)
+			ctx := authentication.SetAuthInfoContext(r.Context(), authInfo)
 			next.ServeHTTP(w, r.WithContext(ctx))
 		case http.StatusUnauthorized:
 			err := authentication.GetAuthController().TerminateSession(r, w)
@@ -91,7 +91,7 @@ func (aHandler AuthenticationHandler) Handle(next http.Handler) http.Handler {
 
 func (aHandler AuthenticationHandler) HandleUnauthenticated(next http.Handler) http.Handler {
 	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		ctx := config.SetAuthInfoContext(r.Context(), &api.AuthInfo{Token: ""})
+		ctx := authentication.SetAuthInfoContext(r.Context(), &api.AuthInfo{Token: ""})
 		next.ServeHTTP(w, r.WithContext(ctx))
 	})
 }
