+++ handlers/utils_test.go
@@ -12,6 +12,7 @@ import (
 	"k8s.io/client-go/tools/clientcmd/api"
 
 	"github.com/kiali/kiali/business"
+	"github.com/kiali/kiali/business/authentication"
 	"github.com/kiali/kiali/config"
 	"github.com/kiali/kiali/kubernetes/kubetest"
 	"github.com/kiali/kiali/prometheus"
@@ -48,7 +49,7 @@ func TestCreateMetricsServiceForNamespace(t *testing.T) {
 	prom, _, _ := utilSetupMocks(t)
 
 	req := httptest.NewRequest("GET", "/foo", nil)
-	req = req.WithContext(config.SetAuthInfoContext(req.Context(), &api.AuthInfo{Token: "test"}))
+	req = req.WithContext(authentication.SetAuthInfoContext(req.Context(), &api.AuthInfo{Token: "test"}))
 
 	w := httptest.NewRecorder()
 	srv, info := createMetricsServiceForNamespace(w, req, prom, "ns1")
@@ -64,7 +65,7 @@ func TestCreateMetricsServiceForNamespaceForbidden(t *testing.T) {
 	prom, _, _ := utilSetupMocks(t)
 
 	req := httptest.NewRequest("GET", "/foo", nil)
-	req = req.WithContext(config.SetAuthInfoContext(req.Context(), &api.AuthInfo{Token: "test"}))
+	req = req.WithContext(authentication.SetAuthInfoContext(req.Context(), &api.AuthInfo{Token: "test"}))
 
 	w := httptest.NewRecorder()
 	srv, info := createMetricsServiceForNamespace(w, req, prom, "nsNil")
@@ -79,7 +80,7 @@ func TestCreateMetricsServiceForSeveralNamespaces(t *testing.T) {
 	prom, _, _ := utilSetupMocks(t)
 
 	req := httptest.NewRequest("GET", "/foo", nil)
-	req = req.WithContext(config.SetAuthInfoContext(req.Context(), &api.AuthInfo{Token: "test"}))
+	req = req.WithContext(authentication.SetAuthInfoContext(req.Context(), &api.AuthInfo{Token: "test"}))
 
 	w := httptest.NewRecorder()
 	srv, info := createMetricsServiceForNamespaces(w, req, prom, []string{"ns1", "ns2", "nsNil"})
